[FOOS-80] Free game fix

+++ src/main/java/pl/ncdchot/foosball/database/repository/RulesRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.repository.CrudRepository;
 import pl.ncdchot.foosball.database.model.Rules;
 
 public interface RulesRepository extends CrudRepository<Rules, Long> {
-	Optional<Rules> findByCreatorIdAndScoreLimitAndTimeLimit(long creatorId, int scoreLimit, long timeLimit);
+	Optional<Rules> findByCreatorIdAndScoreLimitAndTimeLimit(long creatorId, int scoreLimit, int timeLimit);
 }
